# openapi.yaml
...
openapi: 3.0.0
paths:
...
  /v1/customers:
    get:
      description: Cursor-paginated merchant-scoped customer list.
      operationId: CustomersController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of results to return (default 25, max 100)
          schema:
            minimum: 1
            maximum: 100
            default: 25
            type: number
        - name: cursor
          required: false
          in: query
          description: Opaque cursor from a previous list response next_cursor field
          schema:
            type: string
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponseDto'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - ApiKey: []
      summary: List customers
      tags:
        - Customers
  /v1/customers/{id}:
    get:
      operationId: CustomersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Customer publicId (cus_...)
          schema:
            type: string
      responses:
        '200':
          description: Customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseDto'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - ApiKey: []
      summary: Retrieve a customer
      tags:
        - Customers
  /api/developer/api-keys:
    get:
      description: >-
        Returns all active API keys for the Freighter-authenticated merchant (Business.id from ht_dashboard cookie).
        secret_key is always null in list responses.
      operationId: ApiKeysController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - SessionCookie: []
      summary: List API keys
      tags:
        - Developer
    post:
      description: >-
        Generates a new API key. The raw secret_key is returned exactly once and cannot be retrieved again. Store it
        securely immediately.
      operationId: ApiKeysController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - SessionCookie: []
      summary: Create API key
      tags:
        - Developer
  /api/developer/api-keys/{id}/rotate:
    post:
      description: >-
        Immediately revokes the existing key and generates a replacement with the same environment. The new secret_key
        is returned exactly once.
      operationId: ApiKeysController_rotate
      parameters:
        - name: id
          required: true
          in: path
          description: API key publicId (key_...)
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponseDto'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - SessionCookie: []
      summary: Rotate API key
      tags:
        - Developer
  /api/developer/api-keys/{id}/revoke:
    post:
      description: Immediately revokes the key. Any in-flight requests using this key will be rejected within seconds.
      operationId: ApiKeysController_revoke
      parameters:
        - name: id
          required: true
          in: path
          description: API key publicId (key_...)
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponseDto'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - SessionCookie: []
      summary: Revoke API key
      tags:
        - Developer
  /api/developer/customers:
    get:
      operationId: DeveloperCustomersController_findAll
      parameters:
        - name: limit
          required: false
          in: query
          description: Number of results to return (default 25, max 100)
          schema:
            minimum: 1
            maximum: 100
            default: 25
            type: number
        - name: cursor
          required: false
          in: query
          description: Opaque cursor from a previous list response next_cursor field
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponseDto'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - SessionCookie: []
      summary: List customers (dashboard)
      tags:
        - Developer
  /api/developer/customers/{id}:
    get:
      operationId: DeveloperCustomersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Customer publicId (cus_...)
          schema:
            type: string
      responses:
        '200':
          description: Customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseDto'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypertronErrorResponseDto'
      security:
        - SessionCookie: []
      summary: Retrieve a customer (dashboard)
      tags:
        - Developer
...
info:
  title: Hypertron Payments API
  description: >-
    Production-grade stablecoin payment gateway API (HyperTone). Public merchant API under `/v1/*` (API key auth).
    Dashboard control-plane under `/api/developer/*` (Freighter session cookie). Contract reference:
    Payments_API_v1_Schema.md.
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://api.hypertron.xyz
    description: Production
components:
  securitySchemes:
    ApiKey:
      scheme: bearer
      bearerFormat: sk_test_... / sk_live_...
      type: http
      description: Secret API key for /v1 routes
    SessionCookie:
      type: apiKey
      in: cookie
      name: ht_dashboard
      description: Freighter session cookie from hypertron-core-backend (HMAC AUTH_SECRET)
    IdempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Required for POST /v1/payments
    RequestId:
      type: apiKey
      in: header
      name: X-Request-Id
      description: Optional client correlation id
  schemas:
    HypertronErrorBodyDto:
      type: object
      properties:
        type:
          type: string
          example: invalid_request_error
          enum:
            - invalid_request_error
            - authentication_error
            - permission_error
            - resource_missing
            - idempotency_error
            - invalid_state_transition
            - unprocessable_entity
            - rate_limit_error
            - api_error
            - service_unavailable
        code:
          type: string
          example: invalid_amount
        message:
          type: string
          example: amount must be a positive decimal string
        param:
          type: string
          example: amount
          description: Present only for field-level validation errors
        request_id:
          type: string
          example: req_01J...
      required:
        - type
        - code
        - message
        - request_id
    HypertronErrorResponseDto:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/HypertronErrorBodyDto'
      required:
        - error
    CustomerResponseDto:
      type: object
      properties:
        id:
          type: string
          example: cus_01JABC456
        object:
          type: string
          example: customer
          enum:
            - customer
        business_id:
          type: string
          example: clxyz...
        email:
          type: object
          example: customer@example.com
          nullable: true
        name:
          type: object
          example: Ada Lovelace
          nullable: true
        metadata:
          type: object
          example:
            crm_id: crm_123
          additionalProperties: true
        payment_count:
          type: number
          example: 3
        lifetime_value:
          type: string
          example: '250.00'
          description: Decimal string
        lifetime_value_currency:
          type: string
          example: USDC
        last_payment_at:
          type: object
          example: '2026-08-03T12:31:10.000Z'
          nullable: true
        created_at:
          type: string
          example: '2026-07-01T10:00:00.000Z'
        updated_at:
          type: string
          example: '2026-08-03T12:31:10.000Z'
      required:
        - id
        - object
        - business_id
        - metadata
        - payment_count
        - lifetime_value
        - lifetime_value_currency
        - created_at
        - updated_at
    CustomerListResponseDto:
      type: object
      properties:
        object:
          type: string
          example: list
          enum:
            - list
        data:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponseDto'
        has_more:
          type: boolean
          example: false
        next_cursor:
          type: object
          example: null
          nullable: true
      required:
        - object
        - data
        - has_more
    ApiKeyResponseDto:
      type: object
      properties:
        id:
          type: string
          example: key_01J...
        object:
          type: object
          example: api_key
        name:
          type: string
          example: Production server
        environment:
          type: string
          enum:
            - test
            - live
          example: test
        key_prefix:
          type: string
          description: Key prefix for display (e.g. sk_test_)
          example: sk_test_
        last_four:
          type: string
          description: Last four characters of the raw key, for identification
          example: abcd
        secret_key:
          type: object
          description: Full raw key — only returned once at creation/rotation. Null on all other responses.
          example: sk_test_ABC123...
          nullable: true
        active:
          type: boolean
          example: true
        last_used_at:
          type: object
          example: '2024-01-01T00:00:00.000Z'
          nullable: true
        created_at:
          type: string
          example: '2024-01-01T00:00:00.000Z'
        revoked_at:
          type: object
          example: null
          nullable: true
      required:
        - id
        - object
        - name
        - environment
        - key_prefix
        - last_four
        - active
        - created_at
    ApiKeyListResponseDto:
      type: object
      properties:
        object:
          type: object
          example: list
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyResponseDto'
      required:
        - object
        - data
    CreateApiKeyDto:
      type: object
      properties:
        name:
          type: string
          description: Human-readable label for this key
          example: Production server
          minLength: 1
          maxLength: 100
        environment:
          type: string
          description: Key environment — test keys cannot access live payments
          enum:
            - test
            - live
          example: test
      required:
        - name
        - environment
...